Derby-176 (partial)
Automatically handle a generated method exceeding the
byte code limit of 65,535 in limited cases of a method
with no parameters and whose stack depth drops to zero
during code generation. Solution is to generate sub methods
with the same signature and pass control to those methods.
The sub methods are private and hidden to any callers, who
just interact with the single top-level method.


git-svn-id: https://svn.apache.org/repos/asf/incubator/derby/code/trunk@160631 13f79535-47bb-0310-9956-ffa450edef68