Fix Derby-189
I have a new patch for this bug which also fixes the problem you
brought up with sql select * from a.t as X. The fix for this required
change in impl.sql.compile.FromBaseTable's method genResultColList().
I changed the code such that we set the TableDescriptor on the
ColumnDescriptor instance. This TableDescriptor is later used by
ResultColumn.getTableName to get the base table name of the column. In
addition to that, I changed ColumnReference.getSourceTableName and
ColumnReference.getSourceSchemaName so that they don't look at the
user supplied correlation name (if any) to fetch the base table/schema
name.

Patch contributed by Mamta Satoor msatoor@gmail.com


git-svn-id: https://svn.apache.org/repos/asf/incubator/derby/code/trunk@169838 13f79535-47bb-0310-9956-ffa450edef68