committed the following change for ksunithaghm@gmail.com
enhancements to the  TestDurabilityProperty test as a first step  to try to take care of  slow machines where inserts may take longer time than estimated bound in this test.

So if inserts in autocommit mode take more time than the estimated bound, then the test now
--checks if time taken in autocommit on and off are within a close range ( this is because syncs are not happening with derby.system.durability=test mode)
-- if not, then it checks if the time taken to do the insert with derby.system.durability=test mode and without this mode set is within close range, if so it reports a failure with a detailed message explaining the above cases and that the test times the inserts and checks if it within the approximate bounds.

Also note, it will work as expected even if all tests (derbyall)  are running with derby.system.durability=test since this(derby.system.durability) property is explicitly set as needed within the test before booting the database.

The bounds work ok with write cache enabled /disabled on my laptop..  I also tried it out on a slow machine that I have access to and it seems fine. Next, I'll look into adding some kind of acceptable bounds as a percentage of the time taken with this mode set for slower machines.   But if someone notices any issues with the timing on different configurations please post to the list. 



git-svn-id: https://svn.apache.org/repos/asf/incubator/derby/code/trunk@170106 13f79535-47bb-0310-9956-ffa450edef68