Attached is the patch to increase the max possible log files number from 222 -1 to 233 -1. Derby creates log file names in sequential order by incrementing log file number by one each time(like log1.dat ....logN.dat).  id's are not reused, so the last possible log file that can be created currently is  limited to 222 - 1 (4194303). Problem with this limit is, it can be possible hit on systems that creates lot of log records very fast with the default log file size (1MB) ,  for example if  1MB of log is created every second this limit can be hit in (4194303 / (60 * 60 * 24))  48 days.
Attached fix bumps the max possible log file number to 233 -1 by borrowing some
unused bits in the log file number and log file size fields in the log instant.
With the same assumption as above if a 1MB of log file is created for every
second; it will take ((8589934591 / (60 * 60 * 24))/365)
approximately 272 years; Which I think is a reasonable limit.

I believe the new limit is a reasonable one , and not necessary to solve this 
problem by implementing a complicated solution that reuses the ids, at this 
moment.



git-svn-id: https://svn.apache.org/repos/asf/incubator/derby/code/trunk@170686 13f79535-47bb-0310-9956-ffa450edef68