XA end() fails with Connection is closed error
For the client there were multiple problems associated with the tracking
of XA_STATE in the client. The approach for this patch is to remove as
much of the client state as possible.

This patch removes all the client XA state except for
    XA_GLOBAL - In an active global transaction 
    XA_LOCAL - No active global transaction.

For Network Server there was a bug that the connection was not being
reset properly when a connection
reset request came from the client. Changed it so that if an ACCRDB
is sent and a connection has already been established it will reset.




git-svn-id: https://svn.apache.org/repos/asf/incubator/derby/code/trunk@179867 13f79535-47bb-0310-9956-ffa450edef68