Derby-335: Add synonym support to Derby. See the following for more description of the functionality.
Submitted by Satheesh Bandaram (satheesh@sourcery.org)

Couple of changes still pending to complete this feature:

    1) Add registering and checking for Dependencies. These ensure all required objects have been droped before droping synonyms and dropping of synonyms invalidate all required objects.
    2) Add more tests and possible minor code cleanups.

Synonym Functionality:
----------------------
Synonym provides an alternate name for a table or a view that is present in the
same schema or another schema. A synonym can also be created for another
synonym, causing nesting of synonyms. A synonym can be used in SELECT, INSERT,
UPDATE, DELETE or LOCK TABLE statements instead of the original qualified table
or view name. Note that a synonym can be created for a table or a view that
doesn't yet exists. But the target table/view must be present before the
synonym can be used. 

Synonyms are supported by all major database vendors, including Oracle, DB2 and
mySQL. DB2 also allows CREATE ALIAS statement, which does exactly same as
CREATE SYNONYM. Creating aliases instead of synonyms is not supported by Oracle
or mySQL, so I propose that Derby not support creating aliases. Synonyms are not
part of SQL-2003 spec, but is a common-SQL statement among major database
vendors. SQL standard doesn't pay attention to DDLs as much, so I suspect they
skipped synonyms. 

I will be adding two new DDL statements to Derby: 

CREATE SYNONYM <SynonymSchema>.<SynonymName> FOR <TargetSchema>.<TargetName> 
DROP SYNONYM <SynonymSchema>.<SynonymName> 

Synonyms share the same namespace as tables or views. It is not possible to
create a synonym with same name as a table that already exists in the same
schema. Similarly, a table/view can't be created that matches a synonym already
present. 




git-svn-id: https://svn.apache.org/repos/asf/incubator/derby/code/trunk@180459 13f79535-47bb-0310-9956-ffa450edef68