Quick patch submitted based on Kristian Waagan's feedback.  I did not feel a JIRA id was necessary for this fix.
From Kristian:

Today I was setting up the environment to be able to build Derby on a
new machine. I quickly browsed through the BUILDING.txt file, and by
mistake I set both 'j13lib' and 'j14lib' to 'JDK1?_HOME/lib' instead of
'JDK1?_HOME/jre/lib'. After a short moment of temporary confusion, I
realized why I did the mistake. In the section that describes what kind
of slashes to use in path names on Windows, the example value for
'j13lib' is 'c:/jdk1.3.1_15/lib'. This section comes in front of the
sections that actually describe what the values should be.

Although this is a tiny, tiny "improvement", I suggest setting the
example value to 'c:/jdk1.3.1_15/jre/lib' to avoid unnecessary errors
caused by not reading the documentation carefully


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@231524 13f79535-47bb-0310-9956-ffa450edef68