DERBY-353: Return last user specified value for IDENTITY BY DEFAULT columns. The fix changed:
a) Inside the InsertResultSet class getSetAutoincrementValue function the Data Dictionary is accessed to get the current identity value. Which is then updated in the identityVal variable which is then used by the identity_val_local() function to display to the user. 

b) The above process did not happen when there was a user supplied value. The system table was not accessed and the setIdentity() function in the GenericLanguageContext class was not called. 

c) My fix attacked the case when the user supplies a value for the identity column which can be done in the case of Generated by default and calls the setIdentity function to store the value in this case also. 

Thanks to Rick for doing code reviews.

Submitted by V Narayanan (V.Narayanan@Sun.COM)


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@267331 13f79535-47bb-0310-9956-ffa450edef68