DERBY-530 ClientDriver ignores Properties object in connect(String url, Properties connectionProperties) method
Send both the properties specified in the info parameter and those specified in the url to the server in the RDBNAM. user and password attributes will be the exception. user and password will be sent via the standard DRDA mechanism and excluded from the attributes sent with RDBNAM whether specified with the url or info properties. As a result of the combination, the order of attributes sent to the server may be different than originally specified in the URL.

Also added additional driver tests and attrbute tests to checkDriver. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@289227 13f79535-47bb-0310-9956-ffa450edef68