DERBY-557 Client driver gets OutOfMemoryError when re-executing statement without closing ResultSet
Attached a patch which fixes the bug. The patch ensures that the ResultSets associated with a Statement/PreparedStatement are removed from CommitAndRollbackListeners_ in org.apache.derby.client.am.Connection when the statement is re-executed.

I have run derbyall with only one error (wrong length of encryption key, not related to the patch). I have also run the program in the problem description (both with Statement and PreparedStatement), and the memory usage doesn't increase over time.

		Contributed by Knut Anders Hatlen
	


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@289539 13f79535-47bb-0310-9956-ffa450edef68