Fixes DERBY-579 by not storing the timeout value in the GenericPreparedStatement class, just passing it through.
SetQueryTimeoutTest updated, now uses a server-side function to delay
queries. This gives predictability and because of that, the running
time of the test has been significantly reduced. Also, the test now
uses the same tables in multiple statements, in order to hit the
statement cache and check that the timeout affects only the right
statements.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@293585 13f79535-47bb-0310-9956-ffa450edef68