DERBY-424: Ported from TRUNK.
Basically, the problem is that, during the compile phase of views,
the reference to the view gets replaced by the view definition, which causes
us to loose the information that the view might have belonged in SESSION schema.
In order to fix this, during the bind phase in FromList, before the view gets
replaced by its definition, I find out if the view is from SESSION schema, If
yes, then I save this information in FromList and this gets used by FromList
when it is asked if it has any items referencing SESSION schema objects. This
information is again lost during the optimization and generate phase and hence
I moved the check for SESSION schema reference to right after the bind phase in
GenericStatement. If there is a reference to SESSION schema object,
GenericStatement will remove the statement from the cache.

Submitted by Mamta Satoor (msatoor@gmail.com)



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.1@326969 13f79535-47bb-0310-9956-ffa450edef68